+++ script/tool/lib/src/common/core.dart
@@ -103,6 +103,7 @@ final Map<Version, Version> _dartSdkForFlutterSdk = <Version, Version>{
   Version(3, 41, 0): Version(3, 11, 0),
   Version(3, 41, 9): Version(3, 11, 5),
   Version(3, 44, 0): Version(3, 12, 0),
+  Version(3, 47, 0): Version(3, 13, 0),
 };
 
 /// Returns the version of the Dart SDK that shipped with the given Flutter