+++ script/tool/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## NEXT
 
+* Adds the Dart SDK version that shipped with Flutter 3.47.0.
 * Adds support for batch release of pre-1.0 packages.
 
 ## 0.14.4